compress-fits/requirements.txt
Flask==1.1.1
google-cloud-storage
google-cloud-firestore
gunicorn==20.0.4
panoptes-utils>=0.2.6
pytest==5.3.2
